+++ src/main/java/com/twilio/rest/notify/v1/ServiceUpdater.java
@@ -26,6 +26,8 @@ public class ServiceUpdater extends Updater<Service> {
     private String facebookMessengerPageId;
     private String defaultApnNotificationProtocolVersion;
     private String defaultGcmNotificationProtocolVersion;
+    private String fcmCredentialSid;
+    private String defaultFcmNotificationProtocolVersion;
 
     /**
      * Construct a new ServiceUpdater.
@@ -116,6 +118,29 @@ public class ServiceUpdater extends Updater<Service> {
     }
 
     /**
+     * The fcm_credential_sid.
+     * 
+     * @param fcmCredentialSid The fcm_credential_sid
+     * @return this
+     */
+    public ServiceUpdater setFcmCredentialSid(final String fcmCredentialSid) {
+        this.fcmCredentialSid = fcmCredentialSid;
+        return this;
+    }
+
+    /**
+     * The default_fcm_notification_protocol_version.
+     * 
+     * @param defaultFcmNotificationProtocolVersion The
+     *                                              default_fcm_notification_protocol_version
+     * @return this
+     */
+    public ServiceUpdater setDefaultFcmNotificationProtocolVersion(final String defaultFcmNotificationProtocolVersion) {
+        this.defaultFcmNotificationProtocolVersion = defaultFcmNotificationProtocolVersion;
+        return this;
+    }
+
+    /**
      * Make the request to the Twilio API to perform the update.
      * 
      * @param client TwilioRestClient with which to make the request
@@ -186,6 +211,14 @@ public class ServiceUpdater extends Updater<Service> {
         
         if (defaultGcmNotificationProtocolVersion != null) {
             request.addPostParam("DefaultGcmNotificationProtocolVersion", defaultGcmNotificationProtocolVersion);
+        }
+        
+        if (fcmCredentialSid != null) {
+            request.addPostParam("FcmCredentialSid", fcmCredentialSid);
+        }
+        
+        if (defaultFcmNotificationProtocolVersion != null) {
+            request.addPostParam("DefaultFcmNotificationProtocolVersion", defaultFcmNotificationProtocolVersion);
         }
     }
 }